+++ docs/getting_started/jaeger_example.py
@@ -16,7 +16,7 @@
 from opentelemetry import trace
 from opentelemetry.exporter import jaeger
 from opentelemetry.sdk.trace import TracerProvider
-from opentelemetry.sdk.trace.export import BatchExportSpanProcessor
+from opentelemetry.sdk.trace.export import BatchSpanProcessor
 
 trace.set_tracer_provider(TracerProvider())
 
@@ -27,7 +27,7 @@ jaeger_exporter = jaeger.JaegerSpanExporter(
 )
 
 trace.get_tracer_provider().add_span_processor(
-    BatchExportSpanProcessor(jaeger_exporter)
+    BatchSpanProcessor(jaeger_exporter)
 )
 
 tracer = trace.get_tracer(__name__)