+++ exporter/opentelemetry-exporter-otlp/tests/test_otlp_trace_exporter.py
@@ -56,7 +56,7 @@ from opentelemetry.sdk.trace import Status as SDKStatus
 from opentelemetry.sdk.trace import StatusCode as SDKStatusCode
 from opentelemetry.sdk.trace import TracerProvider, _Span
 from opentelemetry.sdk.trace.export import (
-    SimpleExportSpanProcessor,
+    SimpleSpanProcessor,
     SpanExportResult,
 )
 from opentelemetry.sdk.util.instrumentation import InstrumentationInfo
@@ -115,7 +115,7 @@ class TestOTLPSpanExporter(TestCase):
         tracer_provider = TracerProvider()
         self.exporter = OTLPSpanExporter(insecure=True)
         tracer_provider.add_span_processor(
-            SimpleExportSpanProcessor(self.exporter)
+            SimpleSpanProcessor(self.exporter)
         )
         self.tracer = tracer_provider.get_tracer(__name__)