+++ opentelemetry-distro/src/opentelemetry/distro/__init__.py
@@ -30,7 +30,7 @@ from opentelemetry.instrumentation.distro import BaseDistro
 from opentelemetry.sdk.resources import Resource
 from opentelemetry.sdk.trace import TracerProvider
 from opentelemetry.sdk.trace.export import (
-    BatchExportSpanProcessor,
+    BatchSpanProcessor,
     SpanExporter,
 )
 from opentelemetry.sdk.trace.ids_generator import IdsGenerator
@@ -95,7 +95,7 @@ def _init_tracing(
             exporter_args["service_name"] = service_name
 
         provider.add_span_processor(
-            BatchExportSpanProcessor(exporter_class(**exporter_args))
+            BatchSpanProcessor(exporter_class(**exporter_args))
         )