+++ opentelemetry-sdk/tests/performance/resource-usage/trace/profile_resource_usage_simple_export.py
@@ -17,7 +17,7 @@ from unittest.mock import patch
 
 from opentelemetry.exporter.otlp.trace_exporter import OTLPSpanExporter
 from opentelemetry.sdk.trace import TracerProvider, sampling
-from opentelemetry.sdk.trace.export import SimpleExportSpanProcessor
+from opentelemetry.sdk.trace.export import SimpleSpanProcessor
 
 TEST_DURATION_SECONDS = 15
 SPANS_PER_SECOND = 10_000
@@ -31,7 +31,7 @@ class MockTraceServiceStub(object):
 old_stub = OTLPSpanExporter._stub
 OTLPSpanExporter._stub = MockTraceServiceStub
 
-simple_span_processor = SimpleExportSpanProcessor(OTLPSpanExporter())
+simple_span_processor = SimpleSpanProcessor(OTLPSpanExporter())
 tracer = TracerProvider(
     active_span_processor=simple_span_processor, sampler=sampling.DEFAULT_ON,
 ).get_tracer("resource_usage_tracer")